+++ extensions/brave-search-with-results/src/hooks/useHistory.ts
@@ -1,78 +1,64 @@
-import { LocalStorage } from "@raycast/api";
+import { useLocalStorage } from "@raycast/utils";
 import { randomUUID } from "crypto";
-import { useCallback, useEffect, useMemo, useState } from "react";
-
-const STORAGE_KEY = "history";
+import { useCallback } from "react";
 
 export enum Type {
-  Web,
-  Image,
+  Web = "web",
+  Image = "image",
 }
 
 export interface HistoryItem {
   id: string;
   query: string;
-  type: Type;
   date: number;
 }
 
-const get = async () => JSON.parse((await LocalStorage.getItem(STORAGE_KEY)) ?? JSON.stringify([])) as HistoryItem[];
-const set = (history: HistoryItem[]) => LocalStorage.setItem(STORAGE_KEY, JSON.stringify(history));
-
 export default function useHistory(type: Type) {
-  const [isLoading, setIsLoading] = useState(false);
-  const [items, setItems] = useState([] as HistoryItem[]);
+  const { value, setValue, removeValue, isLoading } = useLocalStorage<HistoryItem[]>(type, []);
 
-  const addToHistory = useCallback(
+  const addHistoryItem = useCallback(
     (query: string) => {
       const newItem = {
         id: randomUUID(),
         query: query.toLocaleLowerCase(),
-        type,
         date: Date.now(),
       } as HistoryItem;
 
-      const nextItems = [
-        newItem,
-        ...items.filter((item) => (item.query === newItem.query && item.type === newItem.type) === false),
-      ];
+      // Don't allow modification while the history is still loading
+      if (isLoading || value == null) {
+        console.warn("history is still loading, cannot add new item");
+        return;
+      }
 
-      setItems(nextItems);
-      return set(nextItems);
+      const nextItems = [newItem, ...value.filter((item) => item.query !== newItem.query)];
+
+      setValue(nextItems);
+      return setValue(nextItems);
     },
-    [items, setItems, type],
+    [value, setValue, type],
   );
 
-  const removeFromHistory = useCallback(
+  const removeHistoryItem = useCallback(
     (id: string) => {
-      const nextItems = [...items.filter((item) => item.id !== id)];
+      // Don't allow modification while the history is still loading
+      if (isLoading || value == null) {
+        console.warn("history is still loading, cannot remove item");
+        return;
+      }
 
-      setItems(nextItems);
-      return set(nextItems);
+      const nextItems = [...value.filter((item) => item.id !== id)];
+
+      setValue(nextItems);
+      return setValue(nextItems);
     },
-    [items, setItems],
+    [value, setValue],
   );
 
-  const clearHistory = useCallback(() => {
-    const nextItems = [] as HistoryItem[];
-
-    setItems(nextItems);
-    return set(nextItems);
-  }, [setItems]);
-
-  useEffect(() => {
-    const initialize = async () => {
-      const items = await get();
-      setItems(items);
-      setIsLoading(false);
-    };
-
-    initialize();
-  }, []);
-
-  const filteredItems = useMemo(() => {
-    return items.filter((item) => item.type === type);
-  }, [items, type]);
-
-  return { isLoading, items: filteredItems, addToHistory, removeFromHistory, clearHistory };
+  return {
+    isLoadingHistory: isLoading,
+    historyItems: value ?? [],
+    addHistoryItem,
+    removeHistoryItem,
+    clearHistory: removeValue,
+  };
 }