+++ extensions/brave-search-with-results/src/hooks/useMode.ts
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useState } from "react";
 
 export enum Mode {
   History,
@@ -6,14 +6,6 @@ export enum Mode {
   Search,
 }
 
-export default function useMode(searching: boolean, query: string): Mode {
-  const mode = useMemo(() => {
-    if (searching === false && query.length === 0) return Mode.History;
-    if (searching === false && query.length > 0) return Mode.Suggestions;
-    if (searching) return Mode.Search;
-
-    throw new Error("Invalid mode");
-  }, [searching, query]);
-
-  return mode;
+export default function useMode(initialMode: Mode = Mode.History) {
+  return useState<Mode>(initialMode);
 }